third_party/blink/web_tests/accessibility/slider-thumb-bounds.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<input id="slider" type="range" min=0 max=10 value=0 style="position: absolute; left: 100px; top: 50px; width: 300px; height: 20px;">

<script>
test(function(t) {
  var axSlider = accessibilityController.accessibleElementById("slider");
  assert_equals(axSlider.role, "AXRole: AXSlider");

  var axSliderThumb = axSlider.childAtIndex(0);
  assert_equals(axSliderThumb.role, "AXRole: AXSliderThumb");

  // The slider thumb should be near the beginning of the slider
  assert_true(Math.abs(axSliderThumb.x - axSlider.x) < 5);
  assert_true(Math.abs(axSliderThumb.y - axSlider.y) < 5);

  // The slider should be wide but the thumb should not be
  assert_true(axSlider.width > 250);
  assert_true(axSliderThumb.width < 30);
}, "Test bounds of slider thumb");
</script>

</body>
</html>
